Using `enum`s in bit fields allows the compiler to use byte accesses
That's fine for RAM - but registers are another matter. If the register can only be accessed as a 32-bit value, this optimisation is problematic.

And setting the `enum` to `: unsigned`, sadly, doesn't fix it, since it then complains that full-width values won't "fit" (although they will).

The solution (until GCC fixes it...) is to revert it to `unsigned`, and merely document the type. My convention is to put the expected type in braces `{ }`.